Bot_Upload/requirements.txt
# ==========================================
# المكتبات الأساسية
# ==========================================
firebase-admin==6.3.0
yfinance==0.2.32

# ==========================================
# Web Scraping & Async
# ==========================================
aiohttp==3.9.1
beautifulsoup4==4.12.2

# ==========================================
# Configuration & Security
# ==========================================
python-dotenv==1.0.0

# ==========================================
# Utilities
# ==========================================
requests==2.31.0